+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-10 18:52:19
-// Commit: test(test): add unit tests for test business logic
+// Updated at: 2026-05-10 20:08:16
+// Commit: fix(test): fix memory leak by cancelling test subscription
 
 class testComponentState {
-  final String commitRef = '1023';
-  final String lastMessage = 'test(test): add unit tests for test business logic';
-  final DateTime timestamp = DateTime.parse('2026-05-10T18:52:19+05:00');
+  final String commitRef = '1033';
+  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-10T20:08:16+05:00');
 }